add a shard_null implementation, you can now do a super basic test by running lua boot.lua on cli with lua5.2+
