+++ data/ai/preload/globals.lua
@@ -1,55 +1,26 @@
 -- Created by Tom J Nowell 2010
 -- Shard AI
 if ShardSpringLua then
-	function shard_include( file )
-		if type(file) ~= 'string' then
-			return nil
-		end
-		local gameFile = "luarules/gadgets/ai/" ..  Game.gameShortName .. "/" .. file .. ".lua"
-		local baseFile = "luarules/gadgets/ai/" .. file .. ".lua"
-		local preloadFile = "luarules/gadgets/ai/preload/" .. file .. ".lua"
-		if VFS.FileExists(gameFile) then
-			-- Spring.Echo("got gameFile", gameFile)
-			return VFS.Include(gameFile)
-		elseif VFS.FileExists(baseFile) then
-			-- Spring.Echo("got baseFile", baseFile)
-			return VFS.Include(baseFile)
-		elseif VFS.FileExists(preloadFile) then
-			-- Spring.Echo("got preloadFile", preloadFile)
-			return VFS.Include(preloadFile)
-		end
-	end
+	require "preload/spring_lua/boot"
+elseif game_engine then
+	require "preload/spring_native/boot"
 else
-	function shard_include( file )
-		if type(file) ~= 'string' then
-			return nil
-		end
-		local ok1, mod1 = pcall( require, game_engine:GameName().."/"..file )
-		if ok1 then
-			return mod1
-		else
-			local ok2, mod2 = pcall( require, file )
-			if ok2 then
-				return mod2
-			else
-				game_engine:SendToConsole("require can't load " .. game_engine:GameName().."/"..file .. " error: " .. mod1)
-				game_engine:SendToConsole("require can't load " .. file .. " error: " .. mod2)
-			end
-		end
-	end
+	require "preload/shard_null/boot"
 end
 
-shard_include "hooks"
-shard_include "class"
-shard_include "aibase"
+require "preload/hooks"
+require "preload/class"
+require "preload/aibase"
 
 if ShardSpringLua then
-	shard_include "spring_lua/unit"
-	shard_include "spring_lua/unittype"
-	shard_include "spring_lua/damage"
-	shard_include "spring_lua/feature"
-	shard_include "spring_lua/controlpoint"
+	require "preload/spring_lua/unit"
+	require "preload/spring_lua/unittype"
+	require "preload/spring_lua/damage"
+	require "preload/spring_lua/feature"
+	require "preload/spring_lua/controlpoint"
+elseif game_engine then
+	shard_include "preload/spring_native/unit"
+	shard_include "preload/spring_native/unittype"
 else
-	shard_include "spring_native/unit"
-	shard_include "spring_native/unittype"
+	-- load null objects
 end
