+++ data/ai/preload/shard_null/boot.lua
@@ -1,0 +1,8 @@
+
+
+function shard_include( file )
+	if type(file) ~= 'string' then
+		return nil
+	end
+	require( file )
+end
